requirements.txt
langchain[all]
llama-cpp-python 
flask
lxml
PyYAML
